% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap}
\alias{bootstrap}
\title{Bootstrap Resampling}
\usage{
bootstrap(data)
}
\arguments{
\item{data}{List or data.frame of data to be resampled.}
}
\description{
Draw a random sample with replacement from some \code{data}. Can
be used to resample training data sets in order to construct multiple
decision trees (with different tree structure) given a single training set.
}
\examples{
bootstrap(iris)
bootstrap(iris)
}
