% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest.R
\name{forest}
\alias{forest}
\title{Create Classification 'forest' Object}
\usage{
forest(formula, data, m, b = 5, minsplit = 20,
  minbucket = round(minsplit/3), maxdepth = 30, bsinclude = FALSE)
}
\arguments{
\item{formula}{A \code{formula} containing variable names of response and
predictors.}

\item{data}{data.frame of data to fit the model on.}

\item{m}{Size of predictor subset. If \code{m} is chosen smaller than the
number of predictors, a random subset of the specified predictors will be
used for tree construction.}

\item{b}{Number of bootstrap samples from \code{data} and number of trees in
forest/ensemble.}

\item{minsplit}{Minimum number of observations needed for a split.}

\item{minbucket}{Minimum number of observations that need to be contained
each of the left and right groups after a node was split.}

\item{maxdepth}{Maximum tree depth. The root node has \code{depth = 0}, so
that \code{maxdepth = 3} constrains the tree to 4 levels.}

\item{bsinclude}{Memory intensive! Includes all bootstraps samples in the
\code{forest} object, if set \code{TRUE}. Used for test puporses.}
}
\value{
An object of class \code{forest}, containing the following elements:
An \code{ensemble} holding \code{b} \code{tree} objects. A list of
\code{names} for response and predictor names as well as response classes.
\code{frequencies} holds \code{data}'s response class frequencies.
\code{importance} holds a ranking of predictor importance, measured by the
mean information gain relative to the predictor with the highest mean gain.
\code{control} holds the options as specified in the function call, as well
as \code{random = FALSE} if all predictors were considered for tree
construction. Bootstrap samples are stored in \code{bootstraps} if
\code{bsinclude = TRUE}.
}
\description{
Creates a fitted classification \code{forest} model. If random
subsets of predictors are used for tree construction (via choice of \code{m})
, the resulting model is a "Random Forest".
}
\details{
Constructs one classification \code{tree} per bootstrap sample. For
a total of \code{b} bootstrap samples, all trees are contained in a tree
\code{ensemble} within the generated \code{forest} object.
}
\examples{
# Sample split
data <- iris
set.seed(1337)
rows <- sample.int(nrow(data), size = floor(0.7*nrow(data)), replace = FALSE)
trainData <- data[rows,]
testData <- data[-rows,]
# Create fitted 'forest' object
fit <- forest(Species ~ .,
              data = trainData,
              m = 3,
              b = 5,
              minsplit = 20,
              minbucket = 5,
              maxdepth = 3,
              bsinclude = FALSE)
}
