\docType{data}
\name{ForestInv01}
\alias{ForestInv01}
\title{Forest Inventory Data - Example 01 }
\format{A data frame with 131 rows and 5 variables}
\description{
  Forest inventory data collected in Eucalyptus spp. forest plantations. This is an example of forest inventory data used in \code{\link{poly5Model}}, \code{\link{plotStem2d}} and \code{\link{plotStem3d}} functions   
}
\details{
  \itemize{ 
  \item id. unique number to identify tree information
  \item hi. height (m) of each section where diameter is taken
  \item di. diameter (cm) in hi  
  \item ht. total height (m) of the tree
  \item dbh. diameter (cm) at breast height (e.g 1.30 m)
 }
}
\keyword{datasets}
