% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gain.R
\name{gain}
\alias{gain}
\title{Information Gain}
\usage{
gain(left, right, var, gini)
}
\arguments{
\item{left}{data.frame, group of observations resulting from a binary split.}

\item{right}{data.frame, group of observations resulting from a binary split.}

\item{var}{Variable over which the information gain will be calculated.}

\item{gini}{Gini value before split.}
}
\value{
Returns the pre-split \code{gini} index minus the weighted sum of
Gini indices of post-split \code{left} and \code{right} groups. Weights are
the relative sizes of both left/right groups. Takes values from [0;1], where
values greater than 0 indicate an increase in node purity.
}
\description{
Calculate the information gain from binary splitting data into
\code{left} and \code{right} groups, based on a pre-split \code{gini} index.
}
\examples{
purity <- gini(iris, "Species")
gain(iris[1:89,], iris[90:nrow(iris),], "Species", purity)
}
