% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gini.R
\name{gini}
\alias{gini}
\title{Gini Index}
\usage{
gini(data, var)
}
\arguments{
\item{data}{List or data.frame of data.}

\item{var}{Variable name over which Gini index will be calculated.}
}
\value{
Numeric from [0;1], where 0 indicates fully homogeneous data, 1
indicates completely heterogeneous data. Calculation: Retrieves the relative
frequencies/probabilities of all classes of \code{var} and returns the gini
as 1 minus the sum over all squared relative frequencies.
}
\description{
Calculate the purity of some \code{data} with respect to
variable \code{var}.
}
\details{
Can be used to determine the purity of a decision tree node.
}
\examples{
gini(iris, "Species")
}
