% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance.ensemble.R
\name{importance.forest}
\alias{importance.forest}
\title{Predictor Importance of Fitted 'forest' Objects}
\usage{
\method{importance}{forest}(object, predictors, ...)
}
\arguments{
\item{object}{Object of class \code{forest}.}

\item{predictors}{Set of predictors as specified in the \code{formula}
argument of the constructor \code{forest()}.}

\item{...}{Additional arguments.}
}
\value{
data.frame containing ranked variable importance (highest to lowest
\code{relMeanGain}).
}
\description{
\code{forest}-version of \code{importance.tree()}. Generates a
ranking of predictor variable importance with respect to their mean
information gain relative to the predictor with the highest mean gain.
}
\details{
Ranking of predictor importance is measured by the mean information
gain relative to the predictor with the highest mean gain. Note:
\code{formula} argument (of the constructor \code{tree()}) but randomly
dropped out of consideration due to the choice of \code{m} (random predictor
subset).
}
