% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance.node.R
\name{importance}
\alias{importance}
\title{Predictor Importance}
\usage{
importance(object, ...)
}
\arguments{
\item{object}{Some object.}

\item{...}{Additional arguments.}
}
\description{
Generic function for generating a variable importance ranking.
}
\details{
Defined as generic function: \code{importance(object, ...)} can
be added as method for other classes.
}
