% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misClassErr.R
\name{misClassErr}
\alias{misClassErr}
\title{Misclassification Error Rate}
\usage{
misClassErr(predicted, observed)
}
\arguments{
\item{predicted}{Vector of prediction values.}

\item{observed}{Vector of observed values.}
}
\description{
Calculates the percentage of falsely predicted values.
}
\examples{
data <- iris
set.seed(1337)
rows <- sample.int(nrow(data), size = floor(0.7*nrow(data)), replace = FALSE)
trainData <- data[rows,]
testData <- data[-rows,]
fit <- tree(Species ~ .,
            data = trainData,
            m = 2,
            minsplit = 20,
            minbucket = 5,
            maxdepth = 3)
pred <- predict(fit, testData)
obs <- testData$Species
misClassErr(pred, obs)
}
