% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parenthesize.table.R
\name{parenthesize}
\alias{parenthesize}
\title{Parenthesize}
\usage{
parenthesize(object, ...)
}
\arguments{
\item{object}{An object from which information is extracted.}

\item{...}{Additional arguments.}
}
\description{
Pre-processes information of some \code{object} for printing.
}
\details{
Defined as generic function: \code{parenthesize(object, ...)} can
be added as method for other classes.
}
