% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parenthesize.table.R
\name{parenthesize.table}
\alias{parenthesize.table}
\title{Parenthesize Contents of Table Object}
\usage{
\method{parenthesize}{table}(object, type = c("counts", "names"), ...)
}
\arguments{
\item{object}{Object of class \code{table}.}

\item{type}{Include either class \code{counts} or \code{names}.}

\item{...}{Additional arguments.}
}
\value{
Returns a string of a variable's class counts or names. Counts or
names are separated by backslashed "\". Note: String contains double "\\",
because "\" needs to be escaped as "\\" for printing using \code{cat()}.
String is surrounded by parantheses.
}
\description{
Primarily used for function \code{print.tree()}. Combines
table contents (counts or names) to a string, seperated by backslashes and
enclosed by parantheses. Provides a handy notation of class frequencies and
names to be printed using e.g. \code{cat()}.
}
\examples{
tbl <- table(iris$Species)
printMe <- parenthesize(tbl, type = "names")
cat(printMe)
printMe <- parenthesize(tbl, type = "counts")
cat(printMe)
}
