% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotStem2d.R
\name{plotStem2d}
\alias{plotStem2d}
\title{Create a 2-D stem plot}
\usage{
plotStem2d(hi,di,col)
}
\arguments{
\item{hi, }{vector of measured tree i heights}

\item{di, }{vector of measured tree diameters (di) at i heights}

\item{col, }{stem color, e.g. "chocolate"}
}
\value{
Returns an object of class "lm"
}
\description{
plotStem2d is used to Create a 2-D stem plot
}
\examples{
# Importing forest inventory data
data(taper) # or import a table e.g. taper<-read.table(taper.csv, sep=",", head=T) 

# Subsetting Tree 1
tree1<-subset(taper,taper[,2]==1)
hi<-tree1$sec
di<-tree1$ccc/pi

# Plotting stem 2d
plotStem2d(hi,di, col="chocolate4")
}
\author{
Carlos A. Silva, Samuel P. C. Carvalho, Carine Klauberg Silva and Manoela de O. Rosa
}

