% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotStem3d.R
\name{plotStem3D}
\alias{plotStem3D}
\title{Create a 3-D stem plot}
\usage{
plotStem3D(hi,di,col, alpha, labels,box)
}
\arguments{
\item{hi, }{vector of trees his}

\item{di, }{vector of trees dis}

\item{col, }{stem color, e.g. "chocolate"}

\item{alpha, }{stem transparency. Set a value from 0 to 1}

\item{labels, }{if TRUE, the section heighs will be labed}

\item{box, }{a 3D box will be added to the plot}
}
\description{
plotStem3D is used to Create a 3-D stem plot
}
\examples{
# Importing forest inventory data
fpath<- system.file("extdata", "taper.csv", package="rForest") 
TreeList<-read.table(fpath, sep=",", head=T) # reading data

# Subsetting Tree 1
T1<-subset(TreeList,TreeList[,2]==1)
hi<-T1$Sec
di<-T1$CCC/pi

# Plotting stem 3d
plotStem3D(hi,di,alpha=1,col="green",labels=TRUE,box=TRUE)
aspect3d(0.6,0.5,2)
title3d(zlab = "Height (m)", col="black") # title
}
\author{
Carlos Alberto Silva
}

