% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.forest.R
\name{predict.forest}
\alias{predict.forest}
\title{Prediction for a Fitted 'forest' Object}
\usage{
\method{predict}{forest}(object, data, type = c("aggr", "prob"), ...)
}
\arguments{
\item{object}{\code{forest} object of fitted \code{tree} models.}

\item{data}{Data containing the response variable to be predicted.}

\item{type}{Can be either \code{"aggr"} or \code{"prob"}. For \code{"aggr"},
an aggregate prediction over all \code{tree} objects held by the
\code{forest} will be generated. For \code{"probs"}, a matrix of estimated
prediction probabilities will be calculated for each response class and
observation in \code{data}.}

\item{...}{Additional arguments.}
}
\description{
Generates a vector of prediction values or a matrix of
prediction probabilities for a \code{forest} of fitted \code{tree} objects.
}
\details{
The prediction probability matrix in fact gives a more detailed
picture of class predictions.
}
\examples{
# Sample split
data <- iris
set.seed(1337)
rows <- sample.int(nrow(data), size = floor(0.7*nrow(data)), replace = FALSE)
trainData <- data[rows,]
testData <- data[-rows,]
# Create fitted 'forest' object
fit <- forest(Species ~ .,
              data = trainData,
              m = 3,
              b = 5,
              minsplit = 20,
              minbucket = 5,
              maxdepth = 3,
              bsinclude = FALSE)
# Make prediction (aggregate/mode)
pred <- predict(fit, testData, type = "aggr")
# Make prediction (probabilites)
pred <- predict(fit, testData, type = "prob")
}
