% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.tree.R
\name{predict.tree}
\alias{predict.tree}
\title{Prediction Based on Fitted 'tree' Object}
\usage{
\method{predict}{tree}(object, data, ...)
}
\arguments{
\item{object}{Fitted object of class \code{tree}.}

\item{data}{Data containing the response variable to be predicted.}

\item{...}{Additional arguments.}
}
\value{
A vector of predictions.
}
\description{
Generates a vector of prediction values for some fitted
\code{tree} object.
}
\details{
Supports classification trees only! Generic \code{predict()}
function for class \code{tree}, so that this method will be dispatched if
\code{predict(object, ...)} is passed an object of class \code{tree}.
Internally uses the recursive function \code{singlePred(object)} to classify
single observations by passing the observation down the tree. A final
prediction is generated by looping through \code{data}.
}
\examples{
# Sample split
data <- iris
set.seed(1337)
rows <- sample.int(nrow(data), size = floor(0.7*nrow(data)), replace = FALSE)
trainData <- data[rows,]
testData <- data[-rows,]
# Create fitted 'tree' object
fit <- tree(Species ~ .,
            data = trainData,
            m = 2,
            minsplit = 20,
            minbucket = 5,
            maxdepth = 3)
# Make prediction
pred <- predict(fit, testData)
}
