% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.leaf.R
\name{print.leaf}
\alias{print.leaf}
\title{Print Content of a 'leaf' Objects}
\usage{
\method{print}{leaf}(x, blank = "", ...)
}
\arguments{
\item{x}{Object of class \code{leaf}.}

\item{blank}{Not passed by user. Used internally for recursion to indent
printed text according to a node's/leaf's depth within the sub(tree)
structure.}

\item{...}{Additional arguments passed to or from other methods..}
}
\description{
Prints response variable's class frequencies from observations
inside a \code{node} object, along with the predicted class (mode) and a
prediction error rate.
}
\examples{
# Sample split
data <- iris
set.seed(1337)
rows <- sample.int(nrow(data), size = floor(0.7*nrow(data)), replace = FALSE)
trainData <- data[rows,]
testData <- data[-rows,]
# Create fitted 'tree' object
fit <- tree(Species ~ .,
            data = trainData,
            m = 4,
            minsplit = 20,
            minbucket = 5,
            maxdepth = 3)
# Root's left node is a leaf
fit$root$left
# Alternatively
print(fit$root$left)
}
