% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.node.R
\name{print.node}
\alias{print.node}
\title{Print (Sub)Tree structure of 'node' and 'leaf' Objects}
\usage{
\method{print}{node}(x, blank = "", ...)
}
\arguments{
\item{x}{Object of class \code{node}.}

\item{blank}{Not passed by user. Used internally for recursion to indent
printed text according to a node's/leaf's depth within the sub(tree)
structure.}

\item{...}{Additional arguments passed to or from other methods.}
}
\description{
Prints the (sub)tree structure of \code{node} and \code{leaf}
objects, including split variables, split values, predictions and class
frequencies.
}
\details{
Internally used in \code{print.tree()} for printing the tree
structure of a \code{tree} object. While \code{print.tree()} always prints
the entire tree, \code{print.node()} can be used to also print subtrees.
}
\examples{
# Sample split
data <- iris
set.seed(1337)
rows <- sample.int(nrow(data), size = floor(0.7*nrow(data)), replace = FALSE)
trainData <- data[rows,]
testData <- data[-rows,]
# Create fitted 'tree' object
fit <- tree(Species ~ .,
            data = trainData,
            m = 4,
            minsplit = 20,
            minbucket = 5,
            maxdepth = 3)
# Print right subtree of root
#fit$root$right
# Alternatively
print(fit$root$right)
}
