% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rule.R
\name{rule}
\alias{rule}
\title{Find Best Binary Split Rule}
\usage{
rule(data, response, predictors)
}
\arguments{
\item{data}{List or data.frame of data to be split.}

\item{response}{Response variable from \code{data} over which the information
gain will be calculated.}

\item{predictors}{Vector of predictor variable names to be considered in the
search for the best possible split rule.}
}
\value{
Returns a list with elements \code{var}, \code{val}, \code{type} and
\code{gain}, containing the split rule and the resulting information gain.
}
\description{
Finds a split rule that yields the highest possible information
gain for some \code{data}, \code{response} and a set of \code{predictors}.
}
\details{
Calculates the information gain using \code{gain()} for all values
of a predictor over all predictors. The predictor and value with the highest
gain are chosen as the best possible binary split rule.
}
\examples{
# Two predictors: "Sepal.Width", "Petal.Length"
rule(iris, "Species", c("Sepal.Width", "Petal.Length"))
# All predictors:
rule(iris, "Species", setdiff(colnames(iris), "Species"))
}
