% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setLeaf.R
\name{setLeaf}
\alias{setLeaf}
\title{Set Classifcation Tree Leaf}
\usage{
setLeaf(group, response, depth = 0)
}
\arguments{
\item{group}{Left or right group of observations to which a leaf will be
assigned.}

\item{response}{Response variable as specified in the call of \code{tree()}.}

\item{depth}{Tree level, at which the leaf is set.}
}
\value{
Object of class \code{leaf}, containing elements \code{depth}, a
group of observations \code{obs}, a response \code{prediction} (based on
mode) and \code{predictionErr}, the relative frequency of response classes
other than \code{prediction}.
}
\description{
For internal use in the tree growing function \code{grow()}.
A classification tree leaf is set, if no further splits are possible or
attractive (no further information gain).
}
\details{
In the tree growing process, a group of observations will be
assigned a leaf if 1) not enough observations to further split, 2) the next
split will result in either one or both left/right groups to be too small, 3)
the tree's specified maxdepth is reached or 4) if the next best split does
not yield information gain.
}
