% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setNode.R
\name{setNode}
\alias{setNode}
\title{Set Classifcation Tree Node}
\usage{
setNode(rule, left, right, response, depth = 0)
}
\arguments{
\item{rule}{Best possible binary split rule for this node.}

\item{left}{Left group of observations resulting from the node's split.}

\item{right}{Right group of observations resulting from the node's split.}

\item{response}{Response variable as specified in the call of \code{tree()}.}

\item{depth}{Tree level, at which the node is set.}
}
\value{
Object of class \code{node}, containing the best possible binary
split \code{rule}, the node's \code{depth} as well as \code{left} and
\code{right} groups of observations.
}
\description{
For internal use in the tree growing function \code{grow()}.
A classification tree node is set, if further splits are possible and
attractive (next split yields information gain).
}
\details{
In the tree growing process, a group of observations will be
assigned a node, if the following four conditions are met: 1) Enough
observations to further split, 2) the next split will result in both
left and right groups to be large enough, 3) the tree's specified maxdepth is
not reached and 4) the next best split yields information gain.
}
