% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setSplit.R
\name{setSplit}
\alias{setSplit}
\title{Binary Split Dataset}
\usage{
setSplit(data, splitVar, splitVal)
}
\arguments{
\item{data}{data.frame of data to be split.}

\item{splitVar}{Variable with respect to which \code{data} shall be split.}

\item{splitVal}{Value of \code{splitVar}, at which \code{data} shall be
split.}
}
\value{
List of \code{left} and \code{right} groups of observations.
}
\description{
Splits \code{data} with respect to some variable \code{splitVar}
and some specified value {splitVal}.
}
