% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.forest.R
\name{summary.forest}
\alias{summary.forest}
\title{Summarize a Fitted 'forest' Object}
\usage{
\method{summary}{forest}(object, test = NULL, ...)
}
\arguments{
\item{object}{Fitted object of class \code{forest}.}

\item{test}{data.frame of testing data. Default \code{NULL}, in which case
not prediction will be performed and related details will not be printed.
If the function is passed some testing data, a confusion matrix and a
corresponding classification error rate will be printed additionally.
Version note: As of now only Supports classification trees!}

\item{...}{Additional arguments passed to or from other methods.}
}
\description{
Prints details information on an object of class \code{forest}.
}
\details{
Generic \code{summary(object, ...)} function for class \code{forest}
, so that this method will be dispatched if \code{summary(object, ...)} is
passed an object of class \code{forest}.
}
\examples{
# Sample split
data <- iris
set.seed(1337)
rows <- sample.int(nrow(data), size = floor(0.7*nrow(data)), replace = FALSE)
trainData <- data[rows,]
testData <- data[-rows,]
# Create fitted 'forest' object
fit <- forest(Species ~ .,
              data = trainData,
              m = 3,
              b = 5,
              minsplit = 20,
              minbucket = 5,
              maxdepth = 3,
              bsinclude = FALSE)
# Summarize 'forest' object
summary(fit)
}
