\name{taper}
\alias{taper}
\docType{data}
\title{Taper table}
\description{
A list of individual tree mesurements for taper models
}
\usage{data(taper)}
\format{
  The format is:'table'
}
\details{
The LiDAR-derived-CHM provided as an example dataset is from a longleaf pine forest at Eglin AFB, USA. 
}
\source{
The CHM was generated using Lastools software. The LiDAR data collection was funded by a grant (11-2-1-11) from the Joint Fire Science Program: Data set for fuels, fire behavior, smoke, and fire effects model development and evaluation-the RxCADRE project. R. Ottmar, PI; multiple Co-Is."
}
\references{
Silva et al. 2015
}
\examples{
data(taper)
## head(taper)
}
\keyword{chm}
