% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traverse.node.R
\name{traverse.node}
\alias{traverse.node}
\title{Tree Traversal (Split Rule Extraction)}
\usage{
\method{traverse}{node}(object, ans = NULL, ...)
}
\arguments{
\item{object}{Object of class \code{node}, in most cases the root node of a
\code{tree} object (see examples).}

\item{ans}{Answer variable, left unspecified by the user. Used internally to
return extracted split rules (and to pass them to the next recursive
function call).}

\item{...}{Additional arguments.}
}
\value{
data.frame of split rules from a classification tree.
}
\description{
Recursively traverses through a classification tree in order to
extract information on split rules at each tree node.
}
\details{
\code{traverse.node(object)} and \code{traverse(object)} can both
be used. In the latter case, \code{traverse.node(object)} will be dispatched
if \code{object} is of class \code{node}.
}
\examples{
# Create fitted 'tree' object
fit <- tree(Species ~ .,
            data = iris,
            m = 4,
            minsplit = 20,
            minbucket = 5,
            maxdepth = 3)
# Traverse to retrieve split rules
traverse(fit$root)
}
