% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traverse.node.R
\name{traverse}
\alias{traverse}
\title{Tree Traversal}
\usage{
traverse(object, ...)
}
\arguments{
\item{object}{Some tree-like object.}

\item{...}{Additional arguments.}
}
\description{
Recursive traversal for reading and editing information of
trees.
}
\details{
Defined as generic function: \code{traverse(object, ...)} can
be added as method for other classes.
}
