\docType{data}
\name{ForestInv01}
\alias{ForestInv01}
\title{Example of how object tree in \code{\link{poly5Model}}, \code{\link{plotStem2d}} and \code{\link{plotStem3d}}function should be formated}
\format{A data frame with 131 rows and 5 variables}
\description{
  This dataset includes forest inventory data collected in Eucalyptus spp. forest plantations plantations. All diameters mensures (dbh and di) should rather be
  in centimeters and heights in meters. 
}
\details{
  \itemize{ 
  \item id. unique number to identify tree information
  \item hi. height (m) of each section where diameter is taken
  \item di. diameter (cm) in hi  
  \item ht. total height (m) of the tree
  \item dbh. diameter (cm) at breast height (e.g 1.30 m)
 }
}
\keyword{datasets}
