% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotStem2d.R
\name{plotStem2d}
\alias{plotStem2d}
\title{2D visualization of tree stems}
\usage{
plotStem2d(hi,di,col)
}
\arguments{
\item{hi, }{vector of measured tree i heights}

\item{di, }{vector of measured tree diameters (di) at i heights}

\item{col, }{stem color, e.g. "chocolate"}
}
\value{
Nothing, but outputs a plot
}
\description{
plotStem3d is used to visualize tree stems in 3D
}
\examples{
# Importing forest inventory data
data(ForestInv01) 

# Subsetting Tree 1
tree1<-subset(ForestInv01,ForestInv01[,1]==1)
hi<-tree1$hi
di<-tree1$di

# Plotting stem 2d
plotStem2d(hi,di, col="forestgreen")
}
\author{
Carlos A. Silva
}
