% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotStem2d.R
\name{plotStem2d}
\alias{plotStem2d}
\title{Create a 2-D stem plot}
\usage{
plotStem2d(hi,di,col,box)
}
\arguments{
\item{hi, }{vector of measured tree i heights}

\item{di, }{vector of measured tree diameters (di) at i heights}

\item{col, }{stem color, e.g. "chocolate"}

\item{box, }{if TRUE, a 2D box will be added to the plot}
}
\value{
Returns an object of class "lm"
}
\description{
plotStem2d is used to Create a 2-D stem plot
}
\examples{
# Importing forest inventory data
fpath<- system.file("extdata", "taper.csv", package="rForest") 
TreeList<-read.table(fpath, sep=",", head=T) # reading data

# Subsetting Tree 1
T1<-subset(TreeList,TreeList[,2]==1)
hi<-T1$sec
di<-T1$ccc/pi

# Plotting stem 2d
plotStem2d(hi,di, col="chocolate4",box=TRUE)
}
\author{
Carlos A. Silva, Samuel P. C. Carvalho, Carine Klauberg Silva and Manoela de O. Rosa
}

