% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotStem3d.R
\name{plotStem3d}
\alias{plotStem3d}
\title{3D visualization of tree stems}
\usage{
plotStem3d(hi,di,col,alpha)
}
\arguments{
\item{hi, }{vector of trees his}

\item{di, }{vector of trees dis}

\item{col, }{stem color, e.g. "chocolate"}

\item{alpha, }{stem transparency. Set a value from 0 to 1}
}
\description{
plotStem3d is used to visualize tree stems in 3D
}
\examples{
# Importing forest inventory data
data(ForestInv01) 

# Subsetting Tree 1
tree1<-subset(ForestInv01,ForestInv01[,1]==1)
hi<-tree1$hi
di<-tree1$di

# Plotting stem 3d
plotStem3d(hi,di,alpha=1,col="forestgreen")
}
\author{
Carlos Alberto Silva
}
