% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotStem3d.R
\name{plotStem3d}
\alias{plotStem3d}
\title{Create a 3-D stem plot}
\usage{
plotStem3d(hi,di,col, alpha)
}
\arguments{
\item{hi, }{vector of trees his}

\item{di, }{vector of trees dis}

\item{col, }{stem color, e.g. "chocolate"}

\item{alpha, }{stem transparency. Set a value from 0 to 1}
}
\description{
plotStem3d is used to Create a 3-D stem plot
}
\examples{
# Importing forest inventory data
# Importing forest inventory data
data(taper) # or import a table e.g. taper<-read.table(taper.csv, sep=",", head=T) 

# Subsetting Tree 1
tree1<-subset(taper,taper[,2]==1)
hi<-tree1$sec
di<-tree1$ccc/pi

# Plotting stem 3d
plotStem3d(hi,di,alpha=1,col="green")
require(rgl)
aspect3d(0.3,0.3,1)
title3d(zlab = "Height (m)", col="black") # title
}
\author{
Carlos Alberto Silva
}

